requirements.txt
quart
gunicorn
quart_cors
openai
python-dotenv
uvicorn